Resolve main offenses from RuboCop
Issues from template (especially `Rakefile`) will be resolved later.
