+++ lib/flame/application.rb
@@ -131,7 +131,8 @@ module Flame
 			def default_config_dirs(root_dir:)
 				result = { root_dir: File.realpath(root_dir) }
 				%i[public views config tmp].each do |key|
-					result[:"#{key}_dir"] = proc { File.join(config[:root_dir], key.to_s) }
+					result[:"#{key}_dir"] =
+						proc { File.join(config[:root_dir], key.to_s) }
 				end
 				result
 			end
