+++ lib/flame/controller.rb
@@ -3,65 +3,46 @@
 require 'forwardable'
 require 'gorilla_patch/namespace'
 
+require_relative 'controller/actions'
 require_relative 'controller/path_to'
 
+## Just because of `autoload`
 module Flame
 	autoload :Render, "#{__dir__}/render"
 
 	## Class initialize when Dispatcher found route with it
 	## For new request and response
 	class Controller
-		extend Forwardable
+		extend Actions
 
-		## Shortcut for not-inherited public methods: actions
-		## @return [Array<Symbol>] array of actions (public instance methods)
-		def self.actions
-			public_instance_methods(false)
-		end
+		class << self
+			using GorillaPatch::Inflections
 
-		## Re-define public instance methods (actions) from parent
-		## @param actions [Array<Symbol>] Actions for inheritance
-		## @param exclude [Array<Symbol>] Actions for excluding from inheritance
-		## @example Inherit all parent actions
-		##   class MyController < BaseController
-		##     inherit_actions
-		##   end
-		## @example Inherit certain parent actions
-		##   class MyController < BaseController
-		##     inherit_actions :index, :show
-		##   end
-		## @example Inherit all parent actions exclude certain
-		##   class MyController < BaseController
-		##     inherit_actions exclude: %i[edit update]
-		##   end
-		def self.inherit_actions(actions = superclass.actions, exclude: [])
-			(actions - exclude).each do |public_method|
-				um = superclass.public_instance_method(public_method)
-				define_method public_method, um
+			## Default root path of the controller for requests
+			def default_path
+				modules = underscore.split('/')
+				parts = modules.pop.split('_')
+				parts.shift if parts.first == 'index'
+				parts.pop if %w[controller ctrl].include? parts.last
+				parts = [modules.last] if parts.empty?
+				Flame::Path.merge nil, parts.join('_')
 			end
-		end
 
-		## Re-define public instance method from module
-		## @example Define actions from module in controller
-		##   class MyController < BaseController
-		##     include with_actions Module1
-		##     include with_actions Module2
-		##     ....
-		##   end
-		def self.with_actions(mod, exclude: [])
-			Module.new do
-				@mod = mod
-				@exclude = exclude
+			def refined_http_methods
+				@refined_http_methods ||= []
+			end
 
-				def self.included(ctrl)
-					ctrl.include @mod
+			private
 
-					(@mod.public_instance_methods - @exclude).each do |meth|
-						ctrl.send :define_method, meth, @mod.public_instance_method(meth)
-					end
+			Flame::Router::HTTP_METHODS.each do |http_method|
+				downcased_http_method = http_method.downcase
+				define_method(downcased_http_method) do |action_path, action = nil|
+					refined_http_methods << [downcased_http_method, action_path, action]
 				end
 			end
 		end
+
+		extend Forwardable
 
 		def_delegators(
 			:@dispatcher,
@@ -188,31 +169,6 @@ module Flame
 		def add_controller_class(args)
 			args.unshift(self.class) if args[0].is_a?(Symbol)
 			args.insert(1, :index) if args[0].is_a?(Class) && !args[1].is_a?(Symbol)
-		end
-
-		class << self
-			using GorillaPatch::Inflections
-
-			## Default root path of the controller for requests
-			def default_path
-				modules = underscore.split('/')
-				parts = modules.pop.split('_')
-				parts.shift if parts.first == 'index'
-				parts.pop if %w[controller ctrl].include? parts.last
-				parts = [modules.last] if parts.empty?
-				Flame::Path.merge nil, parts.join('_')
-			end
-
-			def refined_http_methods
-				@refined_http_methods ||= []
-			end
-
-			Flame::Router::HTTP_METHODS.each do |http_method|
-				downcased_http_method = http_method.downcase
-				define_method(downcased_http_method) do |action_path, action = nil|
-					refined_http_methods << [downcased_http_method, action_path, action]
-				end
-			end
 		end
 	end
 end
